requirements.txt
langchain-core 
langchain-openai
python-dotenv
python-frontmatter
